fixed bug user can not add an item again which is already in the wishlist or in the cart

+++ clickMartBackend/Controllers/cart.routes.js
@@ -5,20 +5,17 @@ const { CartModel } = require("../models/cartModel");
 
 const cartController = express.Router();
 
-
 // Getting items from the carts
 
-cartController.get("/all_cart_items",async(req,res)=>{
-    const {userId} = req.body;
-    try {
-        const all_cart_items = await CartModel.find({userId});
-        return res.json({message:"success",items:all_cart_items})
-    } catch (err) {
-        console.log(err);
-    }
-    
-})
-
+cartController.get("/all_cart_items", async (req, res) => {
+  const { userId } = req.body;
+  try {
+    const all_cart_items = await CartModel.find({ userId });
+    return res.json({ message: "success", items: all_cart_items });
+  } catch (err) {
+    console.log(err);
+  }
+});
 
 // Adding items to the cart
 
@@ -39,6 +36,12 @@ cartController.post("/add_item_in_cart", async (req, res) => {
     rating,
     brand,
   } = singleItem;
+
+  const existingItemInCart = await CartModel.findOne({ id: id, userId });
+
+  if (existingItemInCart) {
+    return res.json({ message: "Item already  exist in cart" });
+  }
 
   try {
     cartCount++;
@@ -87,40 +90,30 @@ cartController.patch("/update_item_in_cart", async (req, res) => {
   }
 });
 
-
 // Deleting items form carts
 
 cartController.delete("/delete_item_in_cart", async (req, res) => {
-    const { id, userId } = req.body;
-    if(!id){
-        return re.json({message:"Please select a new item"})
-    }
-  
-    const singleItem = await CartModel.findOne({ id: id, userId: userId });
-    if(!singleItem){
-      return res.json({message:"No such item exists in cart"})
-    }
-    const {title} = singleItem;
-  
-    try {
-      const newItemInCart = await CartModel.findOneAndDelete(
-        { id, userId },
-      );
-  
-      res.json({status:"Success",message:`${title} has removed from your cart list`,});
-    } catch (err) {
-      console.log(err);
-    }
-  });
-  
+  const { id, userId } = req.body;
+  if (!id) {
+    return re.json({ message: "Please select a new item" });
+  }
 
+  const singleItem = await CartModel.findOne({ id: id, userId: userId });
+  if (!singleItem) {
+    return res.json({ message: "No such item exists in cart" });
+  }
+  const { title } = singleItem;
 
+  try {
+    const newItemInCart = await CartModel.findOneAndDelete({ id, userId });
 
-
-
-
-
-
-
+    res.json({
+      status: "Success",
+      message: `${title} has removed from your cart list`,
+    });
+  } catch (err) {
+    console.log(err);
+  }
+});
 
 module.exports = { cartController };
